# Conflicts:
#	src/test/java/com/sysco/app/service/ItemServiceImplTest.java

+++ src/main/java/com/sysco/app/controller/OrderController.java
@@ -2,6 +2,8 @@ package com.sysco.app.controller;
 
 import com.sysco.app.exceptions.EntityNotFoundException;
 import com.sysco.app.exceptions.ErrorCode;
+import com.sysco.app.exceptions.RestExceptionHandler;
+import com.sysco.app.exceptions.ValidationException;
 import com.sysco.app.model.Order;
 import com.sysco.app.service.OrderService;
 import io.swagger.annotations.Api;
@@ -16,12 +18,16 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.Errors;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+
+import javax.validation.ConstraintViolationException;
 import javax.validation.Valid;
+import javax.validation.constraints.Pattern;
 import java.time.Instant;
 import java.util.Date;
 
-
+@Validated
 @RestController
 @RequestMapping(value = "/orders")
 @Api(value = "orders", description = "Operations pertaining to orders in Sysco Order Manager")
@@ -81,8 +87,8 @@ public class OrderController {
             @ApiResponse(code = 404, message = "Order not found")
     })
     @GetMapping(value = "/{id}")
-    public ResponseEntity<Order> getOrderById(@PathVariable("id") String id) {
-
+    public ResponseEntity<Order> getOrderById(@Pattern(regexp = "[0-9a-z]*", message = "Id should be of varchar type")
+                                              @PathVariable("id") String id) {
         Order order = orderService.readOrder(id);
         if(order == null) {
             String errorMessage = "OrderController.getOrderById: Empty order";
@@ -98,10 +104,9 @@ public class OrderController {
 
     @ApiOperation(value = "Update an order for a given Id",produces = "application/json")
     @ApiResponses( value = {
-            @ApiResponse(code = 400, message = "Invalid ID supplied"),
+            @ApiResponse(code = 400, message = "Invalid input"),
             @ApiResponse(code = 401, message = "Authorization failed"),
-            @ApiResponse(code = 404, message = "Order not found"),
-            @ApiResponse(code = 405, message = "Validation exception")
+            @ApiResponse(code = 404, message = "Order not found")
     })
     @PutMapping(value = "/{id}")
     public ResponseEntity<Order> updateOrder(@PathVariable("id") String id, @RequestBody Order order) {
@@ -127,15 +132,22 @@ public class OrderController {
 
     @ApiOperation(value = "Delete an order for a given Id", produces = "application/json")
     @ApiResponses( value = {
-            @ApiResponse(code = 204, message = "No content"),
-            @ApiResponse(code = 400, message = "Invalid ID supplied"),
+            @ApiResponse(code = 204, message = "Successfully processed"),
+            @ApiResponse(code = 400, message = "Invalid input"),
             @ApiResponse(code = 401, message = "Authorization failed"),
-            @ApiResponse(code = 404, message = "Order not found"),
-            @ApiResponse(code = 405, message = "Validation exception")
+            @ApiResponse(code = 404, message = "Order not found")
     })
-    @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
+    @DeleteMapping(value = "/{id}")
     public ResponseEntity<Order> deleteOrder(@PathVariable String id){
         Order currentOrder = orderService.readOrder(id);
+        if(currentOrder==null)
+        {
+            String errorMessage = "OrderController.deleteOrder: Empty order";
+            LOGGER.error(errorMessage);
+            throw new EntityNotFoundException(errorMessage,
+                    ErrorCode.NO_ITEM_FOR_THE_ID, OrderController.class);
+
+        }
         orderService.deleteOrder(id);
         return new ResponseEntity<Order>(currentOrder, HttpStatus.NO_CONTENT);
     }
